+++ index.html
@@ -233,7 +233,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="C:\Users\nasta\Downloads\Игра\2221559-kekstagram-1\js\main.js"></script>
+  <script src="main.js" type="module"></script>
 
 </body>
 </html>
